Продолжаю писать второй запррос
